fix bug with container run on containerd
Partial-Jira-bug: CEM-23526

+++ contrail-analyticsdb/hooks/containerd_engine.py
@@ -342,7 +342,7 @@ class Containerd(container_engine_base.Container):
             f.write(" ".join(args))
         changed |= self._if_changed(run_filename)
 
-        if not changed:
+        if self._if_container_exists(cnt_name) and not changed:
             os.remove(env_filename)
             os.remove(run_filename)
             return
